fix: reuse an existing branch when dispatching a watched task
When `run --watch` dispatched a task whose worktree directory was gone but
whose git branch still existed, `createWorktree` ran `git worktree add -b
<branch>` and crashed with "a branch named '<branch>' already exists". The
dispatcher caught it as a hard failure, so the task re-failed on every tick
and never started.

This is routine: teardown deletes the branch only best-effort
(`deleteBranchBestEffort` swallows failures), and operators sometimes remove
the worktree directory by hand — leaving the branch behind. Eligibility only
treats a task as already-provisioned when the directory is found on disk, so
the leftover branch is never noticed until creation crashes on it.

`createWorktree` now probes for the branch first (`git branch --list`) and,
when it exists, attaches it into the new worktree (`git worktree add <dir>
<branch>`, no `-b`, no fetch) — preserving the branch's prior commits rather
than crashing. The fresh-branch path is unchanged.

The `create` suite was extracted to `worktrees.create.test.ts` because
`worktrees.test.ts` had reached the 2000-line `max-lines` lint cap.
