+++ src/lib/worktrees.test.ts
@@ -101,20 +101,6 @@ function makeUserInfo(username: string): ReturnType<typeof userInfo> {
   return { username, uid: 0, gid: 0, shell: null, homedir: "/tmp" };
 }
 
-function makeBillingWorkdirConfig(projectDir: string): ResolvedConfig {
-  return makeConfig({
-    projectDir,
-    knownRepositories: ["billing"],
-    repositories: [
-      {
-        name: "billing",
-        provision: { create: "create ${dir}", remove: "remove ${dir}" },
-        workdir: "services/api",
-      },
-    ],
-  });
-}
-
 function hasArguments(arguments_: readonly string[], ...needles: readonly string[]): boolean {
   return needles.every((needle) => arguments_.includes(needle));
 }
@@ -328,439 +314,6 @@ describe(findByTask, () => {
     const config = makeConfig({ projectDir });
 
     expect(findByTask(config, "team-99")).toStrictEqual([]);
-  });
-});
-
-describe(create, () => {
-  setupTempProjectDir();
-
-  it("probes origin/HEAD, then fetches and worktree-adds the auto-detected default branch", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    const config = makeConfig({ projectDir });
-    runCommandMock.mockImplementation((_command, arguments_) => {
-      // oxlint-disable-next-line vitest/no-conditional-in-test -- discriminator picks out the symbolic-ref probe so it returns origin/<branch>
-      if (hasArguments(arguments_, "symbolic-ref", "refs/remotes/origin/HEAD")) {
-        return "origin/main\n";
-      }
-      return "";
-    });
-
-    const actual = await create(config, {
-      repository: "repo-a",
-      task: "team-1",
-    });
-
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      [
-        "-C",
-        path.join(projectDir, "repo-a"),
-        "symbolic-ref",
-        "--short",
-        "refs/remotes/origin/HEAD",
-      ],
-      {},
-    );
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      ["-C", path.join(projectDir, "repo-a"), "fetch", "origin", "main"],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      [
-        "-C",
-        path.join(projectDir, "repo-a"),
-        "worktree",
-        "add",
-        "-b",
-        "dev-team-1",
-        path.join(projectDir, "repo-a-team-1"),
-        "origin/main",
-      ],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-    expect(actual.kind).toBe("host");
-    expect(actual.dir).toBe(path.join(projectDir, "repo-a-team-1"));
-  });
-
-  it("creates worktrees for multi-segment source task ids", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    const config = makeConfig({ projectDir });
-    runCommandMock.mockImplementation((_command, arguments_) => {
-      // oxlint-disable-next-line vitest/no-conditional-in-test -- discriminator picks out the symbolic-ref probe so it returns origin/<branch>
-      if (hasArguments(arguments_, "symbolic-ref", "refs/remotes/origin/HEAD")) {
-        return "origin/main\n";
-      }
-      return "";
-    });
-
-    const actual = await create(config, {
-      repository: "repo-a",
-      task: "gc-20260608-001",
-    });
-
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      [
-        "-C",
-        path.join(projectDir, "repo-a"),
-        "worktree",
-        "add",
-        "-b",
-        "dev-gc-20260608-001",
-        path.join(projectDir, "repo-a-gc-20260608-001"),
-        "origin/main",
-      ],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-    expect(actual).toMatchObject({
-      repository: "repo-a",
-      task: "gc-20260608-001",
-      branchName: "dev-gc-20260608-001",
-      dir: path.join(projectDir, "repo-a-gc-20260608-001"),
-      kind: "host",
-    });
-  });
-
-  it("creates worktrees for single-segment todo-txt task ids", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    const config = makeConfig({ projectDir });
-    runCommandMock.mockImplementation((_command, arguments_) => {
-      // oxlint-disable-next-line vitest/no-conditional-in-test -- discriminator picks out the symbolic-ref probe so it returns origin/<branch>
-      if (hasArguments(arguments_, "symbolic-ref", "refs/remotes/origin/HEAD")) {
-        return "origin/main\n";
-      }
-      return "";
-    });
-
-    const actual = await create(config, {
-      repository: "repo-a",
-      task: "rrr",
-    });
-
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      [
-        "-C",
-        path.join(projectDir, "repo-a"),
-        "worktree",
-        "add",
-        "-b",
-        "dev-rrr",
-        path.join(projectDir, "repo-a-rrr"),
-        "origin/main",
-      ],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-    expect(actual).toMatchObject({
-      repository: "repo-a",
-      task: "rrr",
-      branchName: "dev-rrr",
-      dir: path.join(projectDir, "repo-a-rrr"),
-      kind: "host",
-    });
-  });
-
-  it("streams git output (stdio inherit) under verbose", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    const config = makeConfig({ projectDir });
-    runCommandMock.mockImplementation((_command, arguments_) => {
-      // oxlint-disable-next-line vitest/no-conditional-in-test -- discriminator reports an origin/main-backed repo
-      if (hasArguments(arguments_, "symbolic-ref", "refs/remotes/origin/HEAD")) {
-        return "origin/main\n";
-      }
-      return "";
-    });
-    setVerbose(true);
-
-    await create(config, { repository: "repo-a", task: "team-1" });
-
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      ["-C", path.join(projectDir, "repo-a"), "fetch", "origin", "main"],
-      { stdio: "inherit", timeoutMs: 0 },
-    );
-  });
-
-  it("uses the per-repo default branch reported by origin/HEAD (e.g. master)", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    const config = makeConfig({ projectDir });
-    runCommandMock.mockImplementation((_command, arguments_) => {
-      // oxlint-disable-next-line vitest/no-conditional-in-test -- discriminator picks out the symbolic-ref probe so it reports a master-backed repo
-      if (hasArguments(arguments_, "symbolic-ref", "refs/remotes/origin/HEAD")) {
-        return "origin/master\n";
-      }
-      return "";
-    });
-
-    await create(config, {
-      repository: "repo-a",
-      task: "team-1",
-    });
-
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      ["-C", path.join(projectDir, "repo-a"), "fetch", "origin", "master"],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      [
-        "-C",
-        path.join(projectDir, "repo-a"),
-        "worktree",
-        "add",
-        "-b",
-        "dev-team-1",
-        path.join(projectDir, "repo-a-team-1"),
-        "origin/master",
-      ],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-  });
-
-  it("falls back to config.git.defaultBranch when origin/HEAD is not set", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    const config = makeConfig({
-      projectDir,
-      git: { remote: "origin", defaultBranch: "trunk" },
-    });
-    runCommandMock.mockImplementation((_command, arguments_) => {
-      // oxlint-disable-next-line vitest/no-conditional-in-test -- mirrors `git symbolic-ref` exit status 1 when refs/remotes/origin/HEAD is unset
-      if (hasArguments(arguments_, "symbolic-ref", "refs/remotes/origin/HEAD")) {
-        throw new Error(
-          "Command failed: git symbolic-ref --short refs/remotes/origin/HEAD\nExit status: 1",
-        );
-      }
-      return "";
-    });
-
-    await create(config, {
-      repository: "repo-a",
-      task: "team-1",
-    });
-
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      [
-        "-C",
-        path.join(projectDir, "repo-a"),
-        "symbolic-ref",
-        "--short",
-        "refs/remotes/origin/HEAD",
-      ],
-      {},
-    );
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      ["-C", path.join(projectDir, "repo-a"), "fetch", "origin", "trunk"],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "git",
-      [
-        "-C",
-        path.join(projectDir, "repo-a"),
-        "worktree",
-        "add",
-        "-b",
-        "dev-team-1",
-        path.join(projectDir, "repo-a-team-1"),
-        "origin/trunk",
-      ],
-      { stdio: "captured", timeoutMs: 0 },
-    );
-  });
-
-  it("rejects when a host worktree already exists for the same task", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    mkdirSync(path.join(projectDir, "repo-a-team-1"));
-    const config = makeConfig({ projectDir });
-
-    await expect(
-      create(config, {
-        repository: "repo-a",
-        task: "team-1",
-      }),
-    ).rejects.toThrow(/already exists/);
-  });
-
-  it("rejects unknown repositories", async () => {
-    const config = makeConfig({ projectDir });
-
-    await expect(
-      create(config, {
-        repository: "ghost",
-        task: "team-1",
-      }),
-    ).rejects.toThrow(/not in workspace.knownRepositories/);
-  });
-
-  it("throws when the repository directory does not exist", async () => {
-    const config = makeConfig({ projectDir });
-
-    await expect(
-      create(config, {
-        repository: "repo-a",
-        task: "team-1",
-      }),
-    ).rejects.toThrow(/Repository not found/);
-  });
-
-  it.each([
-    ["empty string", ""],
-    ["bare dot", "."],
-    ["double dot", ".."],
-    ["forward slash", "team/123"],
-    ["backslash", String.raw`team\123`],
-    ["embedded ..", "team-..-123"],
-    ["traversal segment", `..${path.sep}evil`],
-    ["wrong shape — uppercase", "TEAM-123"],
-    ["wrong shape — trailing whitespace", "team-123 "],
-    ["wrong shape — leading hyphen", "-rrr"],
-    ["wrong shape — trailing hyphen", "rrr-"],
-  ])("rejects invalid task %s", async (_label, task) => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    const config = makeConfig({ projectDir });
-
-    await expect(
-      create(config, {
-        repository: "repo-a",
-        task,
-      }),
-    ).rejects.toThrow(/must be a plain task id/);
-  });
-
-  it("throws when the OS username is empty", async () => {
-    mkdirSync(path.join(projectDir, "repo-a"));
-    userInfoMock.mockReturnValue(makeUserInfo(""));
-    const config = makeConfig({ projectDir });
-
-    await expect(
-      create(config, {
-        repository: "repo-a",
-        task: "team-1",
-      }),
-    ).rejects.toThrow(/Could not determine OS username/);
-  });
-
-  it("create runs the create template via sh and skips git fetch/worktree add", async () => {
-    userInfoMock.mockReturnValue(makeUserInfo("paul"));
-    runCommandMock.mockReturnValue("");
-
-    const config = makeConfig({
-      projectDir,
-      knownRepositories: ["billing"],
-      repositories: [
-        {
-          name: "billing",
-          provision: {
-            create: "graft new ${branch} billing --from ${baseRef} --dir ${dir}",
-            remove: "graft rm ${branch} -f",
-          },
-        },
-      ],
-    });
-    const controller = new AbortController();
-
-    const entry = await create(
-      config,
-      { repository: "billing", task: "team-220" },
-      controller.signal,
-    );
-
-    expect(entry.dir).toBe(path.join(projectDir, "billing-team-220"));
-    expect(entry.branchName).toBe("paul-team-220");
-
-    // No git fetch / worktree add — only the sh -c template, with the abort
-    // signal forwarded.
-    const commands = runCommandMock.mock.calls.map((call) => call[0]);
-    expect(commands).not.toContain("git");
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "sh",
-      [
-        "-c",
-        `graft new 'paul-team-220' billing --from 'origin/main' --dir '${path.join(projectDir, "billing-team-220")}'`,
-      ],
-      expect.objectContaining({ cwd: projectDir, timeoutMs: 0, signal: controller.signal }),
-    );
-  });
-
-  it("streams create template output (stdio inherit) under verbose", async () => {
-    userInfoMock.mockReturnValue(makeUserInfo("paul"));
-    runCommandMock.mockReturnValue("");
-    const config = makeConfig({
-      projectDir,
-      knownRepositories: ["billing"],
-      repositories: [
-        {
-          name: "billing",
-          provision: { create: "graft new ${branch}", remove: "graft rm ${branch} -f" },
-        },
-      ],
-    });
-    setVerbose(true);
-
-    await create(config, { repository: "billing", task: "team-220" });
-
-    expect(runCommandMock).toHaveBeenCalledWith(
-      "sh",
-      ["-c", "graft new 'paul-team-220'"],
-      expect.objectContaining({ cwd: projectDir, stdio: "inherit", timeoutMs: 0 }),
-    );
-  });
-
-  it("removes the worktree when the configured workdir is missing after create", async () => {
-    runCommandMock.mockReturnValue("");
-    const worktreeDir = path.join(projectDir, "billing-team-220");
-    const config = makeBillingWorkdirConfig(projectDir);
-
-    await expect(create(config, { repository: "billing", task: "team-220" })).rejects.toThrow(
-      /workdir "services\/api" not found/,
-    );
-    expect(runCommandMock.mock.calls.map((call) => call[1][1])).toContain(
-      `remove '${worktreeDir}'`,
-    );
-  });
-
-  it("returns the entry when the configured workdir exists after create", async () => {
-    const config = makeConfig({
-      projectDir,
-      knownRepositories: ["billing"],
-      repositories: [
-        { name: "billing", provision: { create: "true", remove: "true" }, workdir: "services/api" },
-      ],
-    });
-    // The create template materializes the subproject as it runs (after the
-    // worktree-already-exists scan, before the workdir-present assertion).
-    runCommandMock.mockImplementation(() => {
-      mkdirSync(path.join(projectDir, "billing-team-220", "services", "api"), { recursive: true });
-      return "";
-    });
-
-    const entry = await create(config, { repository: "billing", task: "team-220" });
-
-    expect(entry.dir).toBe(path.join(projectDir, "billing-team-220"));
-  });
-
-  it("removes the worktree when the configured workdir is a file after create", async () => {
-    const worktreeDir = path.join(projectDir, "billing-team-220");
-    const config = makeBillingWorkdirConfig(projectDir);
-    runCommandMock
-      .mockImplementationOnce(() => {
-        mkdirSync(path.join(worktreeDir, "services"), { recursive: true });
-        writeFileSync(path.join(worktreeDir, "services", "api"), "");
-        return "";
-      })
-      .mockReturnValue("");
-
-    await expect(create(config, { repository: "billing", task: "team-220" })).rejects.toThrow(
-      /workdir "services\/api" not found/,
-    );
-    expect(runCommandMock.mock.calls.map((call) => call[1][1])).toContain(
-      `remove '${worktreeDir}'`,
-    );
   });
 });
